requirements.txt
# common 
numpy==1.19.4
pyyaml
rospkg
rosnumpy
wind-pypcd
rich
# 
# 
# PyGObject
# opencv-python



